pyproject.toml
[tool.poetry]
name = "fast-api-currency-exchanger"
version = "0.1.0"
description = "API for currency exchanger"
authors = ["IvanGorbunov <falseprogrammerfirst@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
python-dotenv = "^1.0.1"
freecurrencyapi = "^0.1.0"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.16"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.1"
httpx = "^0.26.0"
mypy = "^1.8.0"
pylint = "^3.0.3"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
